backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==1.0.7
clerk-backend-api==2.0.2
click==8.1.8
coloredlogs==15.0.1
cryptography==43.0.3
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.9
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.3
jsonschema-specifications==20250325
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.14
langchain-openai==0.2.14
langchain-text-splitters==0.3.1
langchainhub==0.2.16
langsmith==0.3.14
markdown-it-py==4.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.58.1
orjson==3.11.2
packaging==24.2
pandas==2.2.3
pydantic==2.13.3
pydantic_core==2.29.0
pygments==2.18.0
pyjwt==2.10.0
pyparsing==3.1.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
referencing==0.36.0
regex==2025.4.26
requests==2.32.3
rpds-py==0.18.1
rsa==4.9
safetensors==0.5.2
setuptools==73.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6.1
sqlalchemy==2.0.36
sqlmodel==0.0.20
starlette==0.44.0
tenacity==8.2.3
tiktoken==0.7.0
tokenizers==0.21.3
tqdm==4.67.1
typing_extensions==5.2.0
uc-micro-py==1.0.0
ujson==5.11.0
urllib3==2.3.1
uvicorn==0.32.1
uvloop==0.21.0
websockets==13.1
yarl==1.9.4
zipp==3.19.2
# Enterprise monitoring and performance
prometheus-client==0.21.1
psycopg2-binary==2.9.9
redis==5.0.2
